+++ src/LZ4/LZ4Codec.cs
@@ -272,7 +272,7 @@ namespace LZ4
 
 		/// <summary>Tries to create a specified <seealso cref="ILZ4Service" /> and tests it.</summary>
 		/// <typeparam name="T">Concrete <seealso cref="ILZ4Service" /> type.</typeparam>
-		/// <returns>A service if suceeded or <c>null</c> if it failed.</returns>
+		/// <returns>A service if succeeded or <c>null</c> if it failed.</returns>
 		[MethodImpl(MethodImplOptions.NoInlining)]
 		private static ILZ4Service TryService<T>()
 			where T: ILZ4Service, new()
@@ -281,8 +281,9 @@ namespace LZ4
 			{
 				return AutoTest(new T());
 			}
-			catch
+			catch (Exception)
 			{
+				// I could use Trace here but portable profile does not have Trace
 				return null;
 			}
 		}