+++ src/misc/LZ4.MiniApp/Properties/AssemblyInfo.cs
@@ -1,16 +1,26 @@
 using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("LZ4.Tests.Continuous")]
+[assembly: AssemblyTitle("LZ4.MiniApp")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Milosz Krajewski")]
-[assembly: AssemblyProduct("LZ4.Tests.Continuous")]
-[assembly: AssemblyCopyright("Copyright (c) 2014, Milosz Krajewski")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("LZ4.MiniApp")]
+[assembly: AssemblyCopyright("Copyright ©  2016")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
+
+// Setting ComVisible to false makes the types in this assembly not visible 
+// to COM components.  If you need to access a type in this assembly from 
+// COM, set the ComVisible attribute to true on that type.
+[assembly: ComVisible(false)]
+
+// The following GUID is for the ID of the typelib if this project is exposed to COM
+[assembly: Guid("c6520eac-3c64-4218-984c-acc4e1ba5cd9")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -23,4 +33,4 @@ using System.Reflection;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.9.93")]
+[assembly: AssemblyFileVersion("1.0.10.93")]